상품 옵션 조회 객체 변경 (#20)
