+++ apps/commerce-api/src/main/kotlin/com/loopers/domain/product/ProductItemRepository.kt
@@ -1,5 +1,4 @@
 package com.loopers.domain.product
 
 interface ProductItemRepository {
-    fun findAllByIds(productItemIds: List<Long>): List<ProductItem>
 }
