+++ apps/commerce-api/src/main/kotlin/com/loopers/domain/product/ProductRepository.kt
@@ -5,4 +5,5 @@ import com.loopers.support.enums.sort.ProductSortType
 interface ProductRepository {
     fun findBySortType(brandId: Long?, sortBy: ProductSortType?, offset: Int, limit: Int): List<Product>
     fun findActiveProductById(id: Long): Product?
+    fun findProductItemsByIds(productItemIds: List<Long>): List<ProductItem>
 }
