+++ apps/commerce-api/src/test/kotlin/com/loopers/infrastructure/product/ProductRepositoryImplTest.kt
@@ -1,9 +1,11 @@
 package com.loopers.infrastructure.product
 
 import com.loopers.domain.product.Product
+import com.loopers.domain.product.ProductItem
 import com.loopers.domain.product.ProductRepository
 import com.loopers.domain.product.vo.ProductStatusType
 import com.loopers.fixture.product.ProductFixture
+import com.loopers.fixture.product.ProductItemFixture
 import com.loopers.support.IntegrationTestSupport
 import com.loopers.support.enums.sort.ProductSortType
 import org.assertj.core.api.Assertions.assertThat
@@ -15,6 +17,7 @@ import java.time.LocalDateTime
 class ProductRepositoryImplTest(
     private val cut: ProductRepository,
     private val jpaRepository: JpaProductRepository,
+    private val jpaProductItemRepository: JpaProductItemRepository,
 ) : IntegrationTestSupport() {
 
     @Nested
@@ -186,6 +189,113 @@ class ProductRepositoryImplTest(
 
             // then
             assertThat(actual).isNull()
+        }
+    }
+
+    @Nested
+    inner class `ID 목록으로 상품 아이템 목록을 조회할 때` {
+
+        @Test
+        fun `존재하는 ID 목록으로 조회하면 해당 상품 아이템들을 반환한다`() {
+            // given
+            val product = ProductFixture.`활성 상품 1`.toEntity()
+            val savedProduct = jpaRepository.saveAndFlush(product)
+
+            val productItem1 = ProductItemFixture.`검은색 라지 만원`.toEntity(savedProduct)
+            val productItem2 = ProductItemFixture.`빨간색 라지 만원`.toEntity(savedProduct)
+            val savedItems = jpaProductItemRepository.saveAllAndFlush(listOf(productItem1, productItem2))
+
+            val requestIds = savedItems.map(ProductItem::id)
+
+            // when
+            val actual = cut.findProductItemsByIds(requestIds)
+
+            // then
+            assertThat(actual).hasSize(2)
+                .extracting("name")
+                .containsExactlyInAnyOrder("검은색 라지", "빨간색 라지")
+        }
+
+        @Test
+        fun `존재하지 않는 ID가 포함된 목록으로 조회하면 존재하는 상품 아이템만 반환한다`() {
+            // given
+            val product = ProductFixture.`활성 상품 1`.toEntity()
+            val savedProduct = jpaRepository.saveAndFlush(product)
+
+            val productItem = ProductItemFixture.`검은색 라지 만원`.toEntity(savedProduct)
+            val savedItem = jpaProductItemRepository.saveAndFlush(productItem)
+
+            val requestIds = listOf(savedItem.id, 2L)
+
+            // when
+            val actual = cut.findProductItemsByIds(requestIds)
+
+            // then
+            assertThat(actual).hasSize(1)
+                .extracting("name")
+                .containsExactly("검은색 라지")
+        }
+
+        @Test
+        fun `삭제된 상품 아이템 ID가 포함된 목록으로 조회하면 삭제되지 않은 상품 아이템만 반환한다`() {
+            // given
+            val product = ProductFixture.`활성 상품 1`.toEntity()
+            val savedProduct = jpaRepository.saveAndFlush(product)
+
+            val activeItem = ProductItemFixture.`검은색 라지 만원`.toEntity(savedProduct)
+            val deletedItem = ProductItemFixture.`빨간색 라지 만원`.toEntity(savedProduct).also(ProductItem::delete)
+            val savedItems = jpaProductItemRepository.saveAllAndFlush(listOf(activeItem, deletedItem))
+
+            val requestIds = savedItems.map(ProductItem::id)
+
+            // when
+            val actual = cut.findProductItemsByIds(requestIds)
+
+            // then
+            assertThat(actual).hasSize(1)
+                .extracting("name")
+                .containsExactly("검은색 라지")
+        }
+
+        @Test
+        fun `빈 ID 목록으로 조회하면 빈 리스트를 반환한다`() {
+            // given
+            val product = ProductFixture.`활성 상품 1`.toEntity()
+            val savedProduct = jpaRepository.saveAndFlush(product)
+
+            val productItem1 = ProductItemFixture.`검은색 라지 만원`.toEntity(savedProduct)
+            val productItem2 = ProductItemFixture.`빨간색 라지 만원`.toEntity(savedProduct)
+            jpaProductItemRepository.saveAllAndFlush(listOf(productItem1, productItem2))
+
+            val emptyIds = emptyList<Long>()
+
+            // when
+            val actual = cut.findProductItemsByIds(emptyIds)
+
+            // then
+            assertThat(actual).isEmpty()
+        }
+
+        @Test
+        fun `조회된 상품 아이템에는 Product가 fetch join으로 함께 조회된다`() {
+            // given
+            val product = ProductFixture.`활성 상품 1`.toEntity()
+            val savedProduct = jpaRepository.saveAndFlush(product)
+
+            val productItem = ProductItemFixture.`검은색 라지 만원`.toEntity(savedProduct)
+            val savedItem = jpaProductItemRepository.saveAndFlush(productItem)
+
+            val requestIds = listOf(savedItem.id)
+
+            // when
+            val actual = cut.findProductItemsByIds(requestIds)
+
+            // then
+            assertThat(actual).hasSize(1)
+                .extracting("product.name")
+                .containsExactly(
+                    "활성상품 1",
+                )
         }
     }
 }
